// src/app/models/course-year-fee.ts
export class CourseYearFeeDto{
      courseId: any;
      title: any;
      startDate: any;
      endDate: any;
      seats: any;
      courseTypeName: any;
      eduLevelId: any;
      eduLevel: any;
      years: any;
      medium: any;
      courseFee: any;
      offer: any;
      schedule: any;
}